drm/msm/dsi-staging: fix lane idle check logic for phy v3
Currently in DSI PHY v3, the check for lane idle state
expects the state of unused lanes bit to be zero, which is
not the case because if any particular lane is not used then
that lane state will be set to one which means stop state.
This causes ulps enable to fail as the lane idle check fails.
This change fixes the logic.

Change-Id: I8e7007911c14da0a8c6a79a53cee7338c5a4cafb
Signed-off-by: Sandeep Panda <spanda@codeaurora.org>